Enforce terminal size
When running the MegaLinter from within a terminal, and if its size can
be determined, pass the terminal size to the container (even though no
tty was detected). See https://stackoverflow.com/a/48016366 for more
information. This is to ensure that the output of some linters
(PowerShell) will adapt to the terminal size.

The implementation respects when the relevant variables COLUMNS and
LINES are enforced.